Update the RFC.
This commit updates the RFC to expand the `InstanceAllocator` trait so that an
instance allocator can validate modules being compiled and also control how a
JIT compiler is tuned.

This will be necessary for the pooling instance allocator to fail module
compilation early if the module being compiled cannot be instantiated by the
configured allocator.

This also renamed the "default" instance allocator to the more descriptive
`OnDemandInstanceAllocator`.