RFC: shared host functions.
This RFC proposes to extend the Wasmtime API to allow users to define host
functions in a `Config` that can be shared between multiple engines and
therefore stores.

Today, `Func` is used to define host functions, but it is tied to a `Store`.
If users desire short-lived stores to ensure short-lived instances, host
functions need to be redefined upon every module instantiation.

By defining host functions with a `Config`, instances can be created
without having to redefine the host functions; this will make for faster module
instantiations in scenarios where a module is repeatedly instantiated.